Add assertions
An assertion failure indicates a possible bug in the code, because the
code has resulted in a program state that violates an assumption about
how the code should behave.

As such, adding assertions ensures important variables in the selected
files do not contain null values or indexes out of the bounds of a list.