Add Priority
By adding the priority attribute to Task, Task can implement
Comparable<Task> and their priority can be compared, which enables
features to be added such as the use of priority queue to generate
list of tasks in order of priority.

Users can now view and update the priority of their tasks and decide
what is important at hand.