Remove bugs and Shorten Long Methods
While long methods should often be avoided, the justification for
exceeding the usual limit of 30 lines for the method parseCommand() is
due to a significant number of commands a user can execute. The method
has been shortened from 170 lines to 43 lines, by using SLAP.

This update has also removed the bugs associated with date and time,
which also affects the storage of the tasks when tasks with invalid
dates pass through. Users are highly discouraged from modifying data
directly from the file as that affects the parsing of tasks from storage
to the TaskList. However, in the event of such modifications, an
exception will occur and users will be advised to delete data from their
file to make sure the application works properly.